+++ src/executor/executor.c
@@ -6,131 +6,13 @@
 /*   By: wlin <wlin@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/22 11:46:39 by wlin              #+#    #+#             */
-/*   Updated: 2024/07/27 18:06:40 by wlin             ###   ########.fr       */
+/*   Updated: 2024/07/27 18:44:05 by wlin             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "executor.h"
 #include "macros.h"
-
-void	free_array(char **array)
-{
-	int	i;
-
-	i = 0;
-	if (array == NULL || *array == NULL)
-		return ;
-	while (array[i])
-		free(array[i++]);
-	free(array);
-	array = NULL;
-}
-
-void	execute_command(char *command_path, char **cmd_args, char **envp)
-{
-	char	**result_array_concat = NULL;
-
-	execve(command_path, cmd_args, envp);
-	if (errno == ENOEXEC)
-	{
-		result_array_concat = array_concat("/bin/sh", cmd_args);
-		execve("/bin/sh", result_array_concat, envp);
-        // close(pipe_fd[WR]);
-		// free_array(re sult_array_concat);
-		result_array_concat = NULL;
-	}
-	else if (errno == ENOENT)
-	{
-		if (char_index(cmd_args[0], '/') != INVALID)
-        {
-			perror(cmd_args[0]);
-            // close(pipe_fd[WR]);
-        }
-		else
-		{
-			write(STDERR_FILENO, cmd_args[0], str_size(cmd_args[0]));
-			write(STDERR_FILENO, ": command not found\n", 20);
-            // close(pipe_fd[WR]);
-			exit(127);
-		}
-	}
-	else
-    {
-		perror_and_exit(cmd_args[0], EXIT_FAILURE);
-        // close(pipe_fd[WR]);
-    }
-}
-
-void	redirection_error_handling(char *filename, int fd_in)
-{
-	if ((errno == ENOENT || errno == EACCES)&& fd_in == INVALID)
-		printf("%s: %s\n", filename, strerror(errno));
-}
-
-void	handle_redirection(t_process *process, t_redirect *redirect)
-{
-	if (redirect->type == GREAT)
-	{
-		close(process->fd_out);
-		process->fd_out = open(redirect->filename, O_CREAT | O_TRUNC
-			| O_RDWR, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
-	}
-	else if (redirect->type == GREAT_GREAT)
-	{
-		close(process->fd_out);
-		process->fd_out = open(redirect->filename, O_CREAT | O_APPEND
-			| O_RDWR, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
-	}
-	else if (redirect->type == LESS)
-	{
-		close(process->fd_in);
-		process->fd_in = open(redirect->filename, O_RDONLY);
-	}
-	else if (redirect->type == LESS_LESS)
-	{
-		close(process->fd_in);
-		process->fd_in = read_here_doc(redirect->filename);
-	}
-	redirection_error_handling(redirect->filename, process->fd_in);
-}
-
-int	is_equal(char *s1, char *s2)
-{
-	if (ft_strlen(s1) == ft_strlen(s2))
-	{
-		if (ft_strncmp(s1, s2, ft_strlen(s1)) == 0)
-			return (EXIT_SUCCESS);
-	}
-	return (EXIT_FAILURE);
-}
-
-t_process	init_process(t_commands *cmds, char **envp, int pipe_read_end_prev)
-{
-	t_process	process;
-	t_redirect	*tmp_redirect;
-
-	tmp_redirect = cmds->redirect;
-	process.envp = envp;
-	process.command = cmds->args;
-	if (process.command != NULL)
-		process.cmd_path = find_cmd_path(getenv("PATH"), process.command[0]);
-	process.fd_in = pipe_read_end_prev;
-	if (cmds->next)
-	{
-		if (pipe(process.pipe_fd) == INVALID)
-        	perror_and_exit("pipe", EXIT_FAILURE);
-		process.fd_out = process.pipe_fd[WR];
-	}
-	else
-		process.fd_out = dup(STDOUT_FILENO);
-	while (tmp_redirect)
-	{
-		handle_redirection(&process, tmp_redirect);
-		tmp_redirect = tmp_redirect->next;
-	}
-	return (process);
-}
 
 int get_wait_status(int status)
 {
@@ -159,6 +41,31 @@ void	wait_process(pid_t *pid_array, int num_cmd)
 	}
 }
 
+void	execute_command(char *command_path, char **cmd_args, char **envp)
+{
+	char	**result_array_concat = NULL;
+
+	execve(command_path, cmd_args, envp);
+	if (errno == ENOEXEC)
+	{
+		result_array_concat = array_concat("/bin/sh", cmd_args);
+		execve("/bin/sh", result_array_concat, envp);
+	}
+	else if (errno == ENOENT)
+	{
+		if (char_index(cmd_args[0], '/') != INVALID)
+			perror(cmd_args[0]);
+		else
+		{
+			write(STDERR_FILENO, cmd_args[0], str_size(cmd_args[0]));
+			write(STDERR_FILENO, ": command not found\n", 20);
+			exit(127);
+		}
+	}
+	else
+		perror_and_exit(cmd_args[0], EXIT_FAILURE);
+}
+
 void	execute_all(t_commands *cmds, char **envp)
 {
 	t_commands	*tmp;
@@ -185,7 +92,6 @@ void	execute_all(t_commands *cmds, char **envp)
 			}
 		}
 		tmp = tmp->next;
-		
 		pipe_read_end_prev = process.pipe_fd[RD];
 	}
 	wait_process(pid, num_cmd);
